rename: feature layer의 직관적인 네이밍을 위한 search -> search-api로 폴더 네임 변경
